+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ```log
+0.1.7 - 2025/11/21
+feat: add cypher infra helper
+fix: add missing suffix on record code regex
+
 0.1.6 - 2025/11/17
 feat: add system resource type and id vo
 feat: add more liaison response statuses